Add assertions and new jar file
Assertions added to ensure assumptions of the program holds true during
runtime.

To deal with the case where there is a bug in the code which would break
an assumption about the program.

As a step towards bug free code, let's add assertions to ensure
instantiated variables are not null and deleted tasks are not null.